Throw exception for status to make current user easier.
CreateUser still needs to be of type ActionResult in order to return a 201 Created. The best solution right now might be to find a more convenient way of handling current user. I've tried to simple make the context method ApiControllerBase.GetAuthorizedUser return ActionResult<User>, but this is still cumbersome to manage when using it in an endpoint.

I have found no way to inject the current user easily.