Add post endpoints and revert lazy loading.
Lazy loading has an issue with JSON serializing when returning a single element. GetResubs returned a correctly serialized resubs, whereas GetResub included the entire owner object and a lazyLoading parameter. Below is the only source of this issue I could find, and it does not fix it.

https://stackoverflow.com/questions/60184661/net-core-3-jsonignore-not-working-when-requesting-single-resource

Therefore I've switched back to both eager and explicit loading using DbContext.Entry